requirements.txt
selenium
beautifullsoup4
python-dotenv
pandas